+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/internal/InternalTestUtils.kt
@@ -2,8 +2,6 @@ package net.corda.testing.node.internal
 
 import net.corda.client.rpc.CordaRPCClient
 import net.corda.client.rpc.CordaRPCClientConfiguration
-import net.corda.client.rpc.internal.drainAndShutdown
-import net.corda.client.rpc.internal.shutdownEvent
 import net.corda.core.CordaException
 import net.corda.core.concurrent.CordaFuture
 import net.corda.core.context.InvocationContext
@@ -11,9 +9,6 @@ import net.corda.core.flows.FlowLogic
 import net.corda.core.internal.FlowStateMachine
 import net.corda.core.internal.concurrent.openFuture
 import net.corda.core.internal.times
-import net.corda.core.messaging.CordaRPCOps
-import net.corda.core.messaging.DataFeed
-import net.corda.core.messaging.pendingFlowsCount
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.getOrThrow
 import net.corda.core.utilities.millis
@@ -27,7 +22,6 @@ import net.corda.testing.node.InMemoryMessagingNetwork
 import net.corda.testing.node.User
 import net.corda.testing.node.testContext
 import org.slf4j.LoggerFactory
-import rx.Observable
 import java.net.Socket
 import java.net.SocketException
 import java.time.Duration
@@ -123,23 +117,4 @@ internal interface InternalMockMessagingService : MessagingService {
     fun stop()
 }
 
-fun NodeHandle.drainAndShutdown(user: User = rpcUsers[0], pollingPeriod: Duration = Duration.ofSeconds(1)): Observable<Unit> {
-
-    return rpc.drainAndShutdown()
-}
-
-fun NodeHandle.shutdownEvent(user: User = rpcUsers[0], pollingPeriod: Duration = Duration.ofSeconds(1)): Observable<Unit> {
-
-    return rpcClient().shutdownEvent(user.username, user.password, pollingPeriod)
-}
-
 fun CordaRPCClient.start(user: User) = start(user.username, user.password)
-
-fun NodeHandle.rpcClient(configuration: CordaRPCClientConfiguration = CordaRPCClientConfiguration.default()) = CordaRPCClient(rpcAddress, configuration)
-
-fun CordaRPCOps.pendingFlowsCount(log: (String) -> Unit): DataFeed<Int, Pair<Int, Int>> {
-
-    val pending = pendingFlowsCount()
-    log("Pending flows: ${pending.snapshot}")
-    return pending.copy(updates = pending.updates.doOnNext { (completed, total) -> log("Draining progress: $completed/$total") }.doOnCompleted { log("Draining progress: complete!") })
-}